+++ Shared/Extensions/WriteFreelyModel+APIHandlers.swift
@@ -244,6 +244,7 @@ extension WriteFreelyModel {
                 #if os(macOS)
                 self.selectedPost = cachedPost
                 #endif
+                cachedPost.status = PostStatus.published.rawValue
             }
         } catch {
             self.currentError = AppError.genericError(error.localizedDescription)
@@ -280,7 +281,6 @@ extension WriteFreelyModel {
         cachedPost.postId = fetchedPost.postId
         cachedPost.rtl = fetchedPost.rtl ?? false
         cachedPost.slug = fetchedPost.slug
-        cachedPost.status = PostStatus.published.rawValue
         cachedPost.title = fetchedPost.title ?? ""
         cachedPost.updatedDate = fetchedPost.updatedDate
     }